+++ apps/frontend/jest.setup.ts
@@ -1,5 +1,5 @@
 import "@repo/testing-config/jest.setup.base";
-import { server } from "./src/mocks/server";
+import { server } from "./src/_mocks/server";
 
 beforeAll(() => server.listen());
 afterEach(() => server.resetHandlers());
