+++ apps/frontend/src/_mocks/browser.ts
