+++ apps/frontend/src/store/redux/features/chat/saga/saga.spec.ts
@@ -1,6 +1,6 @@
 import { http } from 'msw';
 import { baseURL } from '@/apis/fetcher';
-import { server } from '@/mocks/server';
+import { server } from '@/_mocks/server';
 import { runSaga } from 'redux-saga';
 import useToastStore from '@repo/store/useToastStore';
